LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedType-AttributeNodes-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes

Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes
